requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
openai==0.27.7
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
